% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.dt2_lib_versions}
\alias{.dt2_lib_versions}
\title{Get all current library versions in DT2}
\usage{
.dt2_lib_versions()
}
\value{
Named character vector
}
\description{
Core library versions here must be kept in sync with
\code{R/dt2_deps.R} and \code{tools/get-dt2-libs.sh}.
Extension versions are read from the registry automatically.
}
\keyword{internal}
