% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.dt2_lock_mode}
\alias{.dt2_lock_mode}
\title{Lock Bootstrap mode per document (prevents mixing core/bootstrap5)}
\usage{
.dt2_lock_mode(bs)
}
\description{
Lock Bootstrap mode per document (prevents mixing core/bootstrap5)
}
\keyword{internal}
