% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.dt2_npm_map}
\alias{.dt2_npm_map}
\title{Mapping of DT2 library names to npm package names}
\usage{
.dt2_npm_map()
}
\description{
Mapping of DT2 library names to npm package names
}
\keyword{internal}
