% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_server_processing.R
\name{.dt2_parse_ssp_request}
\alias{.dt2_parse_ssp_request}
\title{Internal: parse DataTables server-side request (query string to list)}
\usage{
.dt2_parse_ssp_request(req, n_cols)
}
\description{
Internal: parse DataTables server-side request (query string to list)
}
\keyword{internal}
