% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.dt2_shell_var_map}
\alias{.dt2_shell_var_map}
\title{Shell script variable names for each library}
\usage{
.dt2_shell_var_map()
}
\description{
Shell script variable names for each library
}
\keyword{internal}
