% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.npm_latest_version}
\alias{.npm_latest_version}
\title{Fetch latest version of an npm package}
\usage{
.npm_latest_version(pkg)
}
\arguments{
\item{pkg}{npm package name}
}
\value{
version string or NA
}
\description{
Fetch latest version of an npm package
}
\keyword{internal}
