% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.patch_file}
\alias{.patch_file}
\title{Replace a string in a file (first occurrence only)}
\usage{
.patch_file(path, old, new)
}
\arguments{
\item{path}{file path}

\item{old}{string to find (fixed, not regex)}

\item{new}{replacement string}
}
\description{
Replace a string in a file (first occurrence only)
}
\keyword{internal}
