% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_check_updates.R
\name{.print_update_table}
\alias{.print_update_table}
\title{Pretty-print the update check table}
\usage{
.print_update_table(result)
}
\description{
Pretty-print the update check table
}
\keyword{internal}
