% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_html.R
\name{dt2_col_template}
\alias{dt2_col_template}
\title{Simple HTML template per column (replace \code{{{VAL}}})}
\usage{
dt2_col_template(options = list(), col, template)
}
\arguments{
\item{options}{Options list.}

\item{col}{Name or index of target column.}

\item{template}{HTML string with \code{{{VAL}}} placeholder.}
}
\value{
Updated \code{options}.
}
\description{
Simple HTML template per column (replace \code{{{VAL}}})
}
