% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_cols_align}
\alias{dt2_cols_align}
\title{Column align (Bootstrap 5 classes)}
\usage{
dt2_cols_align(options = list(), cols, align = c("left", "center", "right"))
}
\arguments{
\item{options}{Options list.}

\item{cols}{Names or 1-based indices.}

\item{align}{"left","center","right".}
}
\value{
Updated \code{options}.
}
\description{
Column align (Bootstrap 5 classes)
}
