% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_cols_hide}
\alias{dt2_cols_hide}
\title{Hide columns}
\usage{
dt2_cols_hide(options = list(), cols)
}
\arguments{
\item{options}{Options list.}

\item{cols}{Names or 1-based indices.}
}
\value{
Updated \code{options}.
}
\description{
Hide columns
}
