% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_html.R
\name{dt2_cols_html}
\alias{dt2_cols_html}
\title{Allow raw HTML rendering via \code{columns.render}}
\usage{
dt2_cols_html(options = list(), cols, js_render)
}
\arguments{
\item{options}{Options list.}

\item{cols}{Names or 1-based indices.}

\item{js_render}{JS function (via \link[htmlwidgets:JS]{htmlwidgets::JS}) with signature \verb{(data, type, row, meta)}
returning a string of HTML when \code{type == "display"}.}
}
\value{
Updated \code{options}.
}
\description{
Mark columns to render raw HTML using a JS render function.
}
