% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_cols_width}
\alias{dt2_cols_width}
\title{Column widths (CSS)}
\usage{
dt2_cols_width(options = list(), map_named)
}
\arguments{
\item{options}{Options list.}

\item{map_named}{Named character vector: c(Col="120px", ...).}
}
\value{
Updated \code{options}.
}
\description{
Column widths (CSS)
}
