% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_deps.R
\name{dt2_dependency_extensions}
\alias{dt2_dependency_extensions}
\title{Build extension dependencies not already included in dt2.yaml}
\usage{
dt2_dependency_extensions(extensions = NULL)
}
\arguments{
\item{extensions}{Character vector of extension names.}
}
\value{
List of htmlDependency objects.
}
\description{
Build extension dependencies not already included in dt2.yaml
}
\keyword{internal}
