% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_draw}
\alias{dt2_draw}
\title{Redraw the table (proxy)}
\usage{
dt2_draw(proxy)
}
\arguments{
\item{proxy}{\code{\link[=dt2_proxy]{dt2_proxy()}}.}
}
\value{
The \code{proxy} object, returned invisibly.
}
\description{
Redraw the table (proxy)
}
