% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_extensions.R
\name{dt2_extensions}
\alias{dt2_extensions}
\title{List available DataTables extensions}
\usage{
dt2_extensions()
}
\value{
A data.frame with columns \code{name}, \code{version}, \code{dir}.
}
\description{
List available DataTables extensions
}
\examples{
dt2_extensions()
}
