% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_formats.R
\name{dt2_format_time_format}
\alias{dt2_format_time_format}
\title{Format a date/time using DataTables' datetime renderer, with locale}
\usage{
dt2_format_time_format(
  options = list(),
  col_specs,
  from = NULL,
  to = "L",
  locale = "pt-br"
)
}
\arguments{
\item{options}{Options list (returned modified).}

\item{col_specs}{Column names or indices to format.}

\item{from}{Input format (e.g. \code{'YYYY-MM-DDTHH:mm:ssZ'} or \code{NULL} for ISO).}

\item{to}{Output format (e.g. \code{'L LTS'}). See moment.js docs.}

\item{locale}{Locale string, e.g. \code{'pt-br'}.}
}
\value{
Modified \code{options}.
}
\description{
Format a date/time using DataTables' datetime renderer, with locale
}
