% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_formats.R
\name{dt2_format_time_relative}
\alias{dt2_format_time_relative}
\title{Relative time using moment.fromNow(), with locale}
\usage{
dt2_format_time_relative(options = list(), col_specs, locale = "pt-br")
}
\arguments{
\item{options}{list of options (returned updated)}

\item{col_specs}{names or indices to format}

\item{locale}{e.g. "pt-br" (requires moment-with-locales)}
}
\value{
The modified \code{options} list with an updated \code{columnDefs} entry.
}
\description{
Relative time using moment.fromNow(), with locale
}
