% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_language}
\alias{dt2_language}
\title{Language helper (either list or JSON url)}
\usage{
dt2_language(options = list(), lang_list = NULL, lang_url = NULL)
}
\arguments{
\item{options}{Options list.}

\item{lang_list}{Named list of language keys.}

\item{lang_url}{URL to a JSON translation file.}
}
\value{
Updated \code{options}.
}
\description{
Language helper (either list or JSON url)
}
