% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_order}
\alias{dt2_order}
\title{Define initial ordering (option \code{order})}
\usage{
dt2_order(options = list(), ...)
}
\arguments{
\item{options}{Options list.}

\item{...}{Vectors like \code{c(col, "asc"/"desc")}. \code{col} may be name or 1-based index.}
}
\value{
Updated \code{options}.
}
\description{
Define initial ordering (option \code{order})
}
