% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy}
\alias{dt2_proxy}
\title{Create a proxy for a DT2 table}
\usage{
dt2_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Widget id used in \code{\link[=dt2_output]{dt2_output()}}.}

\item{session}{Shiny session.}
}
\value{
A \code{"DT2Proxy"} object.
}
\description{
Create a proxy for a DT2 table
}
