% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy_order}
\alias{dt2_proxy_order}
\title{Order the table (proxy)}
\usage{
dt2_proxy_order(proxy, ..., columns = NULL)
}
\arguments{
\item{proxy}{\code{\link[=dt2_proxy]{dt2_proxy()}}.}

\item{...}{Vectors \code{c(col, "asc"/"desc")}. If you pass \code{columns}, names will be resolved to indices.}

\item{columns}{Optional character vector of column names to resolve names to indices.}
}
\value{
The \code{proxy} object, returned invisibly.
}
\description{
Order the table (proxy)
}
