% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy_page}
\alias{dt2_proxy_page}
\title{Page navigation (proxy)}
\usage{
dt2_proxy_page(
  proxy,
  page = c("first", "previous", "next", "last", "number"),
  number = NULL
)
}
\arguments{
\item{proxy}{A \code{\link[=dt2_proxy]{dt2_proxy()}} object.}

\item{page}{Navigation action: \code{"first"}, \code{"previous"}, \code{"next"}, \code{"last"},
or \code{"number"} (go to a specific page).}

\item{number}{Page number (1-based). Only used when \code{page = "number"}.}
}
\value{
The proxy, invisibly.
}
\description{
Page navigation (proxy)
}
