% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy_search}
\alias{dt2_proxy_search}
\title{Global search (proxy)}
\usage{
dt2_proxy_search(
  proxy,
  value,
  regex = FALSE,
  smart = TRUE,
  caseInsensitive = TRUE
)
}
\arguments{
\item{proxy}{A \code{\link[=dt2_proxy]{dt2_proxy()}} object.}

\item{value}{Search string.}

\item{regex}{Logical; treat \code{value} as a regular expression? Default: \code{FALSE}.}

\item{smart}{Logical; use DataTables smart search? Default: \code{TRUE}.}

\item{caseInsensitive}{Logical; case-insensitive search? Default: \code{TRUE}.}
}
\value{
The proxy, invisibly.
}
\description{
Global search (proxy)
}
