% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_utils.R
\name{dt2_register_renderer}
\alias{dt2_register_renderer}
\title{Register a named JS renderer}
\usage{
dt2_register_renderer(name, js)
}
\arguments{
\item{name}{Unique name (character scalar).}

\item{js}{A \code{htmlwidgets::JS()} function or a JSON helper expression.}
}
\value{
Invisibly, the name.
}
\description{
Register a named JS renderer
}
