% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_replace_data}
\alias{dt2_replace_data}
\title{Replace all data in the table (proxy)}
\usage{
dt2_replace_data(proxy, data)
}
\arguments{
\item{proxy}{\code{\link[=dt2_proxy]{dt2_proxy()}}.}

\item{data}{New data.frame (will be serialized).}
}
\value{
The \code{proxy} object, returned invisibly.
}
\description{
Replace all data in the table (proxy)
}
