% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_search_global}
\alias{dt2_search_global}
\title{Set global search (option \code{search})}
\usage{
dt2_search_global(
  options = list(),
  value,
  regex = FALSE,
  smart = TRUE,
  caseInsensitive = TRUE
)
}
\arguments{
\item{options}{Options list.}

\item{value}{Text.}

\item{regex, smart, caseInsensitive}{Search flags.}
}
\value{
Updated \code{options}.
}
\description{
Set global search (option \code{search})
}
