% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_select_rows}
\alias{dt2_select_rows}
\title{Select rows (proxy; Select extension)}
\usage{
dt2_select_rows(proxy, indexes, reset = TRUE)
}
\arguments{
\item{proxy}{\code{\link[=dt2_proxy]{dt2_proxy()}}.}

\item{indexes}{1-based row indices.}

\item{reset}{If TRUE, clear selection before selecting.}
}
\value{
The \code{proxy} object, returned invisibly.
}
\description{
Select rows (proxy; Select extension)
}
