% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_utils.R
\name{dt2_use_renderer}
\alias{dt2_use_renderer}
\title{Use a named JS renderer on columns}
\usage{
dt2_use_renderer(options = list(), col_specs, name)
}
\arguments{
\item{options}{Options list (returned modified).}

\item{col_specs}{Column names or indices.}

\item{name}{Name used in \code{dt2_register_renderer()}.}
}
\value{
Modified \code{options}.
}
\description{
Use a named JS renderer on columns
}
