% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_cols_escape}
\alias{dt2_cols_escape}
\title{Escape/unescape columns content}
\usage{
dt2_cols_escape(options = list(), cols, escape = TRUE)
}
\arguments{
\item{options}{Options list.}

\item{cols}{Names or indices.}

\item{escape}{If FALSE, tells DT to trust HTML (use with care).}
}
\value{
Updated \code{options}.
}
\description{
Escape/unescape columns content
}
