% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_formats.R
\name{dt2_format_time_format}
\alias{dt2_format_time_format}
\title{Format a date/time using DataTables' datetime renderer, with locale}
\usage{
dt2_format_time_format(
  options = list(),
  col_specs,
  from = NULL,
  to = "L",
  locale = "pt-br"
)
}
\arguments{
\item{from}{input format (e.g. 'YYYY-MM-DDTHH:mm:ssZ' or null for ISO)}

\item{to}{output format (e.g. 'L LTS') - ver moment.js}

\item{locale}{e.g. 'pt-br'}
}
\description{
Format a date/time using DataTables' datetime renderer, with locale
}
