% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_length_menu}
\alias{dt2_length_menu}
\title{Length menu helper}
\usage{
dt2_length_menu(
  options = list(),
  values = c(10, 25, 50, -1),
  labels = c("10", "25", "50", "All")
)
}
\arguments{
\item{options}{Options list.}

\item{values}{Numeric vector of page lengths (e.g., c(10,25,50,-1)).}

\item{labels}{Character vector of labels.}
}
\value{
Updated \code{options}.
}
\description{
Length menu helper
}
