% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_shiny.R
\name{dt2_output}
\alias{dt2_output}
\title{Shiny output for DT2}
\usage{
dt2_output(outputId, width = "100\%", height = "auto")
}
\arguments{
\item{outputId}{Output id.}

\item{width, height}{Size.}
}
\description{
Shiny output for DT2
}
