% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy_page}
\alias{dt2_proxy_page}
\title{Page navigation (proxy)}
\usage{
dt2_proxy_page(
  proxy,
  page = c("first", "previous", "next", "last", "number"),
  number = NULL
)
}
\description{
Page navigation (proxy)
}
