% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_proxy.R
\name{dt2_proxy_search}
\alias{dt2_proxy_search}
\title{Global search (proxy)}
\usage{
dt2_proxy_search(
  proxy,
  value,
  regex = FALSE,
  smart = TRUE,
  caseInsensitive = TRUE
)
}
\description{
Global search (proxy)
}
