% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_server_processing.R
\name{dt2_ssp_handler}
\alias{dt2_ssp_handler}
\title{Default server-side handler (filter/order/page)}
\usage{
dt2_ssp_handler(names)
}
\arguments{
\item{names}{character() column names in display order.}
}
\value{
function(data, req) -> list(draw, recordsTotal, recordsFiltered, data)
}
\description{
Default server-side handler (filter/order/page)
}
