% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_shiny.R
\name{dt2_state}
\alias{dt2_state}
\title{Access the current state snapshot of a DT2 table}
\usage{
dt2_state(input, id)
}
\arguments{
\item{input}{Shiny input.}

\item{id}{Widget id.}
}
\value{
A list with \code{reason}, \code{order}, \code{search}, \code{page}, \code{selected}, \code{state}.
}
\description{
Access the current state snapshot of a DT2 table
}
