% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_theme}
\alias{dt2_theme}
\title{Bootstrap 5 table theme toggles}
\usage{
dt2_theme(options = list(), compact = TRUE, striped = TRUE, hover = TRUE)
}
\arguments{
\item{options}{Options list.}

\item{compact, striped, hover}{Toggles for table classes.}
}
\value{
Updated \code{options}.
}
\description{
Bootstrap 5 table theme toggles
}
