% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_options.R
\name{dt2_use_buttons}
\alias{dt2_use_buttons}
\title{Enable Buttons (extension) and define buttons}
\usage{
dt2_use_buttons(options = list(), buttons = c("copy", "csv", "excel", "print"))
}
\arguments{
\item{options}{Options list.}

\item{buttons}{Vector of button ids (e.g., c("copy","csv","excel","print","colvis")).}
}
\value{
Updated \code{options}.
}
\description{
Enable Buttons (extension) and define buttons
}
