% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2_shiny.R
\name{render_dt2}
\alias{render_dt2}
\title{Shiny render for DT2}
\usage{
render_dt2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression returning \code{\link[=dt2]{dt2()}}.}

\item{env, quoted}{Standard shinyRenderWidget args.}
}
\description{
Shiny render for DT2
}
